requirements.txt
python-gitlab==3.3.*
PyYAML==6.0
psycopg2-binary==2.9.1
google-api-python-client==2.42.*
oauth2client==4.1.3
num2words==0.5.9
WooCommerce==3.0.0
python-dateutil==2.8.0
cryptography==2.4.2
paramiko==2.4.2
filelock==3.0.12
pytz
retrying
ruamel.yaml
jinja2==3.1.*
mergedeep
